GC: SharePoint: Helper function consolidation. (#1782)
## Description
Move helper functions & test structs to `helper_test.go`
<!-- Insert PR description-->

## Type of change

- [x] :hamster: Trivial/Minor


## Test Plan

- [x] :zap: Unit test
