+++ src/internal/connector/sharepoint/helper_test.go
@@ -1,12 +1,21 @@
 package sharepoint
 
 import (
+	"testing"
+
 	msgraphsdk "github.com/microsoftgraph/msgraph-sdk-go"
 	"github.com/pkg/errors"
+	"github.com/stretchr/testify/require"
 
 	"github.com/alcionai/corso/src/internal/connector/graph"
+	"github.com/alcionai/corso/src/internal/connector/onedrive"
 	"github.com/alcionai/corso/src/pkg/account"
 )
+
+// ---------------------------------------------------------------------------
+// SharePoint Test Services
+// ---------------------------------------------------------------------------
+type MockGraphService struct{}
 
 type testService struct {
 	client      msgraphsdk.GraphServiceClient
@@ -15,8 +24,20 @@ type testService struct {
 }
 
 //------------------------------------------------------------
-// Functions to comply with graph.Service Interface
+// Interface Functions: @See graph.Service
 //------------------------------------------------------------
+
+func (ms *MockGraphService) Client() *msgraphsdk.GraphServiceClient {
+	return nil
+}
+
+func (ms *MockGraphService) Adapter() *msgraphsdk.GraphRequestAdapter {
+	return nil
+}
+
+func (ms *MockGraphService) ErrPolicy() bool {
+	return false
+}
 
 func (ts *testService) Client() *msgraphsdk.GraphServiceClient {
 	return &ts.client
@@ -29,6 +50,10 @@ func (ts *testService) Adapter() *msgraphsdk.GraphRequestAdapter {
 func (ts *testService) ErrPolicy() bool {
 	return false
 }
+
+// ---------------------------------------------------------------------------
+// Helper Functions
+// ---------------------------------------------------------------------------
 
 func createTestService(credentials account.M365Config) (*testService, error) {
 	{
@@ -49,4 +74,17 @@ func createTestService(credentials account.M365Config) (*testService, error) {
 
 		return &service, nil
 	}
+}
+
+func expectedPathAsSlice(t *testing.T, tenant, user string, rest ...string) []string {
+	res := make([]string, 0, len(rest))
+
+	for _, r := range rest {
+		p, err := onedrive.GetCanonicalPath(r, tenant, user, onedrive.SharePointSource)
+		require.NoError(t, err)
+
+		res = append(res, p.String())
+	}
+
+	return res
 }
